header.php -
Replace navigation with static site menu and connected to WP Theme menu

Needed to connect bootstrap and while there added google fonts and HTML5 shiv declaration

Fixed nav bar img logo path to connect to image in assets

screenshot.php -

Replaced image with Mike Cramer Eagle image to be displayed with Theme image

style.css -

Copied custom.css static site styles at the bottom of this file as this is the file that  WP uses instead. Removed styles in this file that will be replaced with bootstrap files and the styles copied from custom.css. Fixed path disconnect to images. Added a rule for .logged-in to adjust the top to display the menu. Finally added a couple of @media rules to adjust the menu for screen resizing.